+++ src/super_gradients/training/utils/predict/prediction_pose_estimation_results.py
@@ -7,7 +7,7 @@ import numpy as np
 from super_gradients.training.utils.predict import ImagePrediction, ImagesPredictions, VideoPredictions, PoseEstimationPrediction
 from super_gradients.training.utils.media.image import show_image, save_image
 from super_gradients.training.utils.media.video import show_video_from_frames, save_video
-from super_gradients.training.utils.visualization.poses import draw_skeleton
+from super_gradients.training.utils.visualization.pose_estimation import draw_skeleton
 
 
 @dataclass
